+++ src/domain/scenes/weapons/sniper/sniper.h
@@ -7,8 +7,8 @@ namespace Game
   class Sniper : public Weapon
   {
   private:
-    const float spriteWidth = 160;
-    const float spriteHeight = 40;
+    const int spriteWidth = 160;
+    const int spriteHeight = 40;
     const SoundEffect projectileSFX = Global::adaptersInstance.audioManager->initSoundEffect("assets/sounds/projectile.wav");
 
     AnimatedSprite shootAnimation = AnimatedSprite({0.01f,
